@@ -4,32 +4,30 @@
 	"version": "0.0.0",
 	"type": "module",
 	"scripts": {
-		"dev": "vite",
-		"build": "tsc && vite build",
-		"dev": "vite --port 3000",
-		"lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-		"preview": "vite preview"
+		"build": "rsbuild build",
+		"check": "biome check --write",
+		"dev": "rsbuild dev --open",
+		"format": "biome format --write",
+		"preview": "rsbuild preview"
 	},
 	"dependencies": {
+		"react": "^18.3.1",
+		"react-dom": "^18.3.1",
 		"date-fns": "^4.1.0",
-		"date-holidays": "^3.19.2",
-		"framer-motion": "^10.12.16",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-hook-form": "^7.44.2",
-		"react-router-dom": "^6.11.2",
+		"date-holidays": "^3.23.12",
+		"framer-motion": "^11.11.8",
+		"react-hook-form": "^7.53.0",
+		"react-router-dom": "^6.26.2",
 		"tabler-icons-react": "^1.56.0"
 	},
 	"devDependencies": {
-		"@biomejs/biome": "^1.9.3",
-		"@types/react": "^18.0.37",
-		"@types/react-dom": "^18.0.11",
-		"@vitejs/plugin-react-swc": "^3.0.0",
-		"autoprefixer": "^10.4.14",
-		"postcss": "^8.4.24",
-		"prettier": "^2.8.8",
-		"tailwindcss": "^3.3.2",
-		"typescript": "^5.0.2",
-		"vite": "^4.3.9"
+		"@biomejs/biome": "^1.8.3",
+		"@rsbuild/core": "^1.0.1",
+		"@rsbuild/plugin-react": "^1.0.1",
+		"@types/react": "^18.3.5",
+		"@types/react-dom": "^18.3.0",
+		"autoprefixer": "^10.4.20",
+		"tailwindcss": "^3.4.13",
+		"typescript": "^5.5.2"
 	}
 }
