@@ -6,27 +6,31 @@ import "./index.css";
 import { IndexPage } from "./pages/IndexPage.tsx";
 
 const SalaryCalculatorPage = React.lazy(() =>
-	import("./pages/salaryCalculator/SalaryCalculatorPage.tsx").then((module) => ({
-		default: module.SalaryCalculatorPage,
-	}))
+	import("./pages/salaryCalculator/SalaryCalculatorPage.tsx").then(
+		(module) => ({
+			default: module.SalaryCalculatorPage,
+		}),
+	),
 );
 
 const FuelCalculatorPage = React.lazy(() =>
 	import("./pages/fuelCalculator/FuelCalculator.tsx").then((module) => ({
 		default: module.FuelCalculatorPage,
-	}))
+	})),
 );
 
 const PpiCalculatorPage = React.lazy(() =>
 	import("./pages/ppiCalculator.tsx").then((module) => ({
 		default: module.PpiCalculatorPage,
-	}))
+	})),
 );
 
 const HolidayCalculator = React.lazy(() =>
-	import("./pages/holidayCalculator/holidayCalculator.tsx").then((module) => ({
-		default: module.HolidayCalculator,
-	}))
+	import("./pages/holidayCalculator/holidayCalculator.tsx").then(
+		(module) => ({
+			default: module.HolidayCalculator,
+		}),
+	),
 );
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
@@ -35,12 +39,24 @@ ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
 			<BrowserRouter>
 				<Routes>
 					<Route index element={<IndexPage />} />
-					<Route path="/calculators/salary" element={<SalaryCalculatorPage />} />
-					<Route path="/calculators/fuel" element={<FuelCalculatorPage />} />
-					<Route path="/calculators/ppi" element={<PpiCalculatorPage />} />
-					<Route path="/calculators/holidays" element={<HolidayCalculator />} />
+					<Route
+						path="/calculators/salary"
+						element={<SalaryCalculatorPage />}
+					/>
+					<Route
+						path="/calculators/fuel"
+						element={<FuelCalculatorPage />}
+					/>
+					<Route
+						path="/calculators/ppi"
+						element={<PpiCalculatorPage />}
+					/>
+					<Route
+						path="/calculators/holidays"
+						element={<HolidayCalculator />}
+					/>
 				</Routes>
 			</BrowserRouter>
 		</Suspense>
-	</React.StrictMode>
+	</React.StrictMode>,
 );
