@@ -1,4 +1,3 @@
-
 import { format } from "date-fns";
 import { classNames } from "../../../../utils/classNames";
 import { formatCurrency, formatNumber } from "../../../../utils/formatNumber";
@@ -12,8 +11,10 @@ type Props = {
 export const MonthInfo = ({ monthWithDetails }: Props) => {
 	const { highestWorkdaysMonth, lowestWorkdaysMonth } = useSalaryContext();
 
-	const isHighestMonth = highestWorkdaysMonth?.formattedDate === monthWithDetails.formattedDate;
-	const isLowestMonth = lowestWorkdaysMonth?.formattedDate === monthWithDetails.formattedDate;
+	const isHighestMonth =
+		highestWorkdaysMonth?.formattedDate === monthWithDetails.formattedDate;
+	const isLowestMonth =
+		lowestWorkdaysMonth?.formattedDate === monthWithDetails.formattedDate;
 
 	return (
 		<div className="flex flex-col gap-2">
@@ -21,7 +22,7 @@ export const MonthInfo = ({ monthWithDetails }: Props) => {
 				className={classNames(
 					"w-max text-lg font-bold",
 					isHighestMonth && "text-green-500",
-					isLowestMonth && "text-yellow-500"
+					isLowestMonth && "text-yellow-500",
 				)}
 			>
 				{format(monthWithDetails.date, "MMMM yyyy")}
