@@ -1,7 +1,16 @@
-
-import { eachMonthOfInterval, endOfYear, isSameMonth, isSameYear, startOfYear } from "date-fns";
+import {
+	eachMonthOfInterval,
+	endOfYear,
+	isSameMonth,
+	isSameYear,
+	startOfYear,
+} from "date-fns";
 import { formatNumber } from "../../../utils/formatNumber";
-import { DayWithDetails, MonthWithDetails, formatMonth } from "../salaryTypes";
+import {
+	type DayWithDetails,
+	type MonthWithDetails,
+	formatMonth,
+} from "../salaryTypes";
 
 type Props = {
 	selectedMonth: Date;
@@ -9,7 +18,11 @@ type Props = {
 	daysWithDetails: Map<string, DayWithDetails>;
 };
 
-export const getMonthsWithDetails = ({ daysWithDetails, hourlyPay, selectedMonth }: Props) => {
+export const getMonthsWithDetails = ({
+	daysWithDetails,
+	hourlyPay,
+	selectedMonth,
+}: Props) => {
 	const months = eachMonthOfInterval({
 		start: startOfYear(selectedMonth),
 		end: endOfYear(selectedMonth),
@@ -17,14 +30,19 @@ export const getMonthsWithDetails = ({ daysWithDetails, hourlyPay, selectedMonth
 
 	return new Map([
 		...(months.map((month) => {
-			const monthsDaysWithDetails = [...daysWithDetails.values()].filter((day) =>
-				isSameMonth(day.date, month)
+			const monthsDaysWithDetails = [...daysWithDetails.values()].filter(
+				(day) => isSameMonth(day.date, month),
 			);
 
 			const isMonthInSelectedYear = isSameYear(month, selectedMonth);
 			const isMonthSelected = isSameMonth(month, selectedMonth);
-			const workdays = monthsDaysWithDetails.filter((day) => day.isWorkday);
-			const workhours = monthsDaysWithDetails.reduce((acc, day) => acc + day.workhours, 0);
+			const workdays = monthsDaysWithDetails.filter(
+				(day) => day.isWorkday,
+			);
+			const workhours = monthsDaysWithDetails.reduce(
+				(acc, day) => acc + day.workhours,
+				0,
+			);
 			const formattedWorkhours = formatNumber(workhours);
 
 			const salary = workhours * hourlyPay;
